Now possible to add a new attribute to a selected node
In the node details section a new button is shown to add an attribute.
By specifying a key and value for that attribute, once selecting
"apply", the attribute will be added to the DOM. An error will be
thrown if trying to add an attribute with an existing key.
When rendering the d3 tree, the root node will no longer be in the
horizontal center but placed at an 8th of the horizontal size in order
to reveal more of the tree from the get go.